Moved ResponsiveTable and Table into Content from where its used
