+++ src/ui/RealmBrowser/Content/Table/HeaderCell.tsx
@@ -18,11 +18,12 @@
 
 import * as classNames from 'classnames';
 import * as React from 'react';
-import { ControlPosition, DraggableCore, DraggableData } from 'react-draggable';
+import { DraggableCore, DraggableData } from 'react-draggable';
 import * as Realm from 'realm';
 
+import { IPropertyWithName } from '../..';
+
 import { ISorting, SortClickHandler } from '.';
-import { IPropertyWithName } from '..';
 
 // This constant should match the $realm-browser-header-handle-width in scss
 const HANDLE_WIDTH = 5;
