+++ src/ui/RealmBrowser/Content/Table/types/StringCell/StringCell.tsx
